requirements.txt
supabase-py==2.3.0
python-dotenv==1.0.0
anthropic==0.18.1
argparse==1.4.0
typing-extensions==4.9.0 